polls/permissions.py
from rest_framework import permissions


class ReadOnly(permissions.BasePermission):

    def has_permission(self, request, view):
        return request.method in permissions.SAFE_METHODS

    def has_object_permission(self, request, view, obj):
        if obj.is_active() is True:
            return request.method in permissions.SAFE_METHODS
