requirements.txt
alembic==1.7.3
APScheduler==3.6.3
argon2-cffi==21.1.0
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==36.0.0
cycler==0.10.0
DateTime==4.3
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==2.2.10
django-allauth==0.47.0
django-filter==21.1
djangorestframework==3.12.4
entrypoints==0.3
ephem==4.0.0.2
flake8==4.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.9.0
mpmath==1.2.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==3.7.1
pytest-sugar==0.9.1
python-dateutil==2.8.2
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2020.5
pywin32==301
pywinpty==1.1.4
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.25
sqlparse==0.4.1
sympy==1.8
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.1.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zope.interface==5.4.0
